[add]admin側の実装１回目
categoriesが連動するため未実装 rubocop済

+++ db/migrate/20200413112724_create_categories.rb
@@ -2,7 +2,7 @@ class CreateCategories < ActiveRecord::Migration[5.2]
   def change
     create_table :categories do |t|
       t.string  :category, null: false
-      t.decimal :effective_life, null: false
+      t.integer :effective_life, null: false
 
       t.timestamps
     end
